Correctly clear session data
