+++ core-bundle/src/Resources/contao/forms/Form.php
@@ -527,6 +527,9 @@ class Form extends Hybrid
 			$_SESSION['FORM_DATA'][$key] = $this->allowTags ? \Input::postHtml($key, true) : \Input::post($key, true);
 		}
 
+		// Store the submitted time to invalidate the session later on
+		$_SESSION['FORM_DATA']['SUBMITTED_AT'] = time();
+
 		$arrFiles = $_SESSION['FILES'];
 
 		// HOOK: process form data callback
